requirements.txt
pillow
ffmpeg-python
numpy
typing